requirements.txt
asgiref==3.4.1
Django==3.2.9
django-quill-editor==0.1.22
django-translated-fields==0.11.3
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
